# roaring: speed-up intersectionCountArrayBitmap