@CLAassistant 2018-03-15
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/pilosa/pilosa?pullRequest=1167) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/pilosa/pilosa?pullRequest=1167) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/pilosa/pilosa?pullRequest=1167) it.</sub>